fixes plain value type being assigned to nullable variables in some scenarios (#251)
